Refactoring: update packages
