+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:clipToPadding="false"
-    tools:context=".MainActivity">
+    tools:context=".ui.MainActivity">
 
     <View
         android:id="@+id/viewDecoration"
